Extracting submission details to a separate component to be able to add features to it
