+++ apps/endatix-hub/app/(main)/forms/[formId]/submissions/[submissionId]/page.tsx
@@ -1,16 +1,13 @@
-import AnswerViewer from '@/features/submissions/ui/answers/answer-viewer';
 import { SubmissionHeader } from '@/features/submissions/ui/details/submission-header';
-import { SubmissionProperties } from '@/features/submissions/ui/details/submission-properties';
 import { SubmissionTopNav } from '@/features/submissions/ui/details/submission-top-nav';
 import { getSubmissionDetailsUseCase } from '@/features/submissions/use-cases/get-submission-details.use-case';
 import { Result } from '@/lib/result';
-import { Model } from 'survey-core';
 import { SectionTitle } from '@/components/headings/section-title';
-import { BackToSubmissionsButton } from '@/features/submissions/ui/details/back-to-submissions-button';
 import { Suspense } from 'react';
 import { Skeleton } from '@/components/ui/skeleton';
 import { pdf } from '@react-pdf/renderer';
 import { SubmissionDataPdf } from '@/components/export/submission-data-pdf';
+import SubmissionDetails from '@/features/submissions/ui/details/submission-details';
 
 type Params = {
   params: Promise<{
@@ -65,69 +62,8 @@ export default async function SubmissionPage({ params, searchParams }: Params) {
       <SubmissionTopNav formId={formId} />
       <SubmissionHeader submissionId={submissionId} formId={formId} />
       <Suspense fallback={<SubmissionDataSkeleton />}>
-        <SubmissionData formId={formId} submissionId={submissionId} />
+        <SubmissionDetails formId={formId} submissionId={submissionId} />
       </Suspense>
-    </>
-  );
-}
-
-async function SubmissionData({
-  formId,
-  submissionId,
-}: {
-  formId: string;
-  submissionId: string;
-}) {
-  const submissionResult = await getSubmissionDetailsUseCase({
-    formId,
-    submissionId,
-  });
-
-  if (Result.isError(submissionResult)) {
-    return (
-      <div>
-        <h1>Submission not found</h1>
-        <BackToSubmissionsButton
-          formId={formId}
-          text="All form submissions"
-          variant="default"
-        />
-      </div>
-    );
-  }
-  const submission = submissionResult.value;
-  if (!submission.formDefinition) {
-    return <div>Form definition not found</div>;
-  }
-  const json = JSON.parse(submission.formDefinition.jsonData);
-  const surveyModel = new Model(json);
-
-  let submissionData = {};
-  try {
-    submissionData = JSON.parse(submission?.jsonData);
-  } catch (ex) {
-    console.warn("Error while parsing submission's JSON data", ex);
-  }
-
-  surveyModel.data = submissionData;
-  const questions = surveyModel.getAllQuestions(false, false, true);
-
-  return (
-    <>
-      <SubmissionProperties submission={submission} />
-      <SectionTitle title="Submission Answers" headingClassName="py-2 my-0" />
-      <div className="grid gap-4">
-        {questions?.map((question) => {
-          return (
-            <div
-              key={question.id}
-              className="grid grid-cols-5 items-center gap-4 mb-6"
-            >
-              <AnswerViewer key={question.id} forQuestion={question} />
-            </div>
-          );
-        })}
-      </div>
     </>
   );
 }
